requirements.txt
# Kirby Super Star Sprite Editor Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
Pillow>=9.0.0
numpy>=1.20.0
PyQt6>=6.0.0

# Testing dependencies (optional)
pytest>=7.0.0
pytest-qt>=4.0.0

# Documentation dependencies (optional)
# sphinx>=4.0.0
# sphinx-rtd-theme>=1.0.0